# packages/server/src/swagger/swagger.yaml
basePath: null
definitions:
...
    Result:
        description: ""
        properties:
            ok:
                type: boolean
                description: ""
        type: object
        required:
            - ok
...
    KeyValueDto:
        description: ""
        properties:
            label:
                type: string
                description: ""
            value:
                type: string
                description: ""
        type: object
        required:
            - label
            - value
    Query:
        description: ""
        properties:
            size:
                type: number
                format: double
                description: ""
            page:
                type: number
                format: double
                description: ""
        type: object
        required:
            - size
            - page
...
    User:
        description: ""
        properties:
            id:
                type: string
                description: ""
            username:
                type: string
                description: ""
            password:
                type: string
                description: ""
            name:
                type: string
                description: ""
            keyword:
                type: string
                description: ""
            avatar:
                type: string
                description: ""
            type:
                type: string
                description: ""
            groups:
                type: array
                items:
                    type: string
                description: ""
            roles:
                type: array
                items:
                    type: string
                description: ""
            email:
                type: string
                description: ""
            mobile:
                type: string
                description: ""
            profile:
                type: object
                description: ""
            isDisable:
                type: boolean
                description: ""
            isAdmin:
                type: boolean
                description: ""
            isApproved:
                type: boolean
                description: ""
            secret:
                type: string
                description: ""
            expired:
                type: number
                format: double
                description: ""
        type: object
        required:
            - id
            - username
            - name
            - avatar
            - type
            - groups
            - roles
            - email
            - mobile
            - isDisable
            - isAdmin
            - isApproved
            - secret
            - expired
    UserRes:
        description: ""
        properties:
            username:
                type: string
                description: ""
            avatar:
                type: string
                description: ""
            email:
                type: string
                description: ""
            name:
                type: string
                description: ""
            mobile:
                type: string
                description: ""
            isAdmin:
                type: boolean
                description: ""
            isApproved:
                type: boolean
                description: ""
            expired:
                type: number
                format: double
                description: ""
            company:
                type: string
                description: ""
            siteUrl:
                type: string
                description: ""
            address:
                type: string
                description: ""
        type: object
        required:
            - username
            - avatar
            - email
            - name
            - mobile
            - isAdmin
            - isApproved
            - expired
    CreateUserReq:
        description: ""
        properties:
            username:
                type: string
                description: ""
            password:
                type: string
                description: ""
            avatar:
                type: string
                description: ""
            email:
                type: string
                description: ""
            name:
                type: string
                description: ""
            mobile:
                type: string
                description: ""
            isAdmin:
                type: boolean
                description: ""
            isApproved:
                type: boolean
                description: ""
            expired:
                type: number
                format: double
                description: ""
            company:
                type: string
                description: ""
            siteUrl:
                type: string
                description: ""
            address:
                type: string
                description: ""
        type: object
        required:
            - username
            - password
            - avatar
            - email
            - name
            - mobile
            - isAdmin
            - isApproved
            - expired
    UsersOfRole:
        description: ""
        properties:
            role:
                type: string
                description: ""
            userIds:
                type: array
                items:
                    type: string
                description: ""
        type: object
        required:
            - role
            - userIds
    EditUserReq:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            mobile:
                type: string
                description: ""
            password:
                type: string
                description: ""
            roles:
                type: array
                items:
                    type: string
                description: ""
            email:
                type: string
                description: ""
            company:
                type: string
                description: ""
            siteUrl:
                type: string
                description: ""
            address:
                type: string
                description: ""
        type: object
        required:
            - id
            - name
    EditProfileReq:
        description: ""
        properties:
            name:
                type: string
                description: ""
            mobile:
                type: string
                description: ""
            password:
                type: string
                description: ""
            roles:
                type: array
                items:
                    type: string
                description: ""
            email:
                type: string
                description: ""
            company:
                type: string
                description: ""
            siteUrl:
                type: string
                description: ""
            address:
                type: string
                description: ""
        type: object
        required:
            - name
    ResultList$User:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/User'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
...
    CreateRoleReq:
        description: ""
        properties:
            name:
                type: string
                description: ""
            description:
                type: string
                description: ""
            permissions:
                type: array
                items:
                    type: string
                description: ""
        type: object
        required:
            - name
            - description
            - permissions
    Role:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            role:
                type: string
                description: ""
            description:
                type: string
                description: ""
            permissions:
                type: array
                items:
                    type: string
                description: ""
        type: object
        required:
            - id
            - name
            - role
            - description
            - permissions
    EditRoleReq:
        description: ""
        properties:
            id:
                type: string
                description: ""
            name:
                type: string
                description: ""
            description:
                type: string
                description: ""
            permissions:
                type: array
                items:
                    type: string
                description: ""
        type: object
        required:
            - id
            - name
            - description
            - permissions
    ResultList$Role:
        description: ""
        properties:
            list:
                type: array
                items:
                    $ref: '#/definitions/Role'
                description: ""
            count:
                type: number
                format: double
                description: ""
            query:
                $ref: '#/definitions/Query'
        type: object
        required:
            - list
...
info:
    description: 项目标准接口
    license:
        name: MIT
    title: 'rest api interface'
    version: 0.0.1
paths:
...
    /user/profile:
        get:
            operationId: UsersProfile
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/User'
            description: ""
            tags:
                - core
            parameters: []
        put:
            operationId: UsersUpdateProfile
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/UserRes'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: user
                    required: true
                    schema:
                        $ref: '#/definitions/EditProfileReq'
    /user/:
        post:
            operationId: UsersCreate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/UserRes'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: user
                    required: true
                    schema:
                        $ref: '#/definitions/CreateUserReq'
        put:
            operationId: UsersUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/UserRes'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: user
                    required: true
                    schema:
                        $ref: '#/definitions/EditUserReq'
    /user/role:
        post:
            operationId: UsersAddUsersToRole
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Result'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: users
                    required: true
                    schema:
                        $ref: '#/definitions/UsersOfRole'
        delete:
            operationId: UsersRemoveAccountFromRole
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Result'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: role
                    required: true
                    type: string
                -
                    description: ""
                    in: query
                    name: id
                    required: true
                    type: string
    /user/search:
        get:
            operationId: UsersSearch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/KeyValueDto'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: value
                    required: false
                    type: string
    /user/query:
        get:
            operationId: UsersQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$User'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
    '/user/{id}':
        get:
            operationId: UsersFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/User'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
...
    /role/:
        post:
            operationId: RolesCreate
            produces:
                - text/html
            responses:
                '204':
                    description: 'No content'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/CreateRoleReq'
        put:
            operationId: RolesUpdate
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Role'
            description: ""
            tags:
                - core
            consumes:
                - application/json
            parameters:
                -
                    description: ""
                    in: body
                    name: entry
                    required: true
                    schema:
                        $ref: '#/definitions/EditRoleReq'
    /role/search:
        get:
            operationId: RolesSearch
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        type: array
                        items:
                            $ref: '#/definitions/KeyValueDto'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: value
                    required: false
                    type: string
    /role/query:
        get:
            operationId: RolesQuery
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/ResultList$Role'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: query
                    name: keyword
                    required: false
                    type: string
                -
                    description: ""
                    in: query
                    name: page
                    required: false
                    type: number
                    format: double
                    default: 1
                -
                    description: ""
                    in: query
                    name: size
                    required: false
                    type: number
                    format: double
                    default: 10
    '/role/{id}':
        get:
            operationId: RolesFindOne
            produces:
                - application/json
            responses:
                '200':
                    description: Ok
                    schema:
                        $ref: '#/definitions/Role'
            description: ""
            tags:
                - core
            parameters:
                -
                    description: ""
                    in: path
                    name: id
                    required: true
                    type: string
...
swagger: '2.0'
securityDefinitions: {}
produces:
    - application/json